changed prime.h to take 64 bit int

+++ interpreter/prime.h
@@ -1,2 +1,4 @@
-int is_prime(const int x);
-int next_prime(int x);
+#include <inttypes.h>
+
+int is_prime(const int64_t x);
+int next_prime(int64_t x);
